Fix minimum image to choose consisitent r_ij
For degenerate points in the cell (points with the same |r_ij| in several cells) the Smith method and Bruteforce method were choosing different r_ij (even though |r_ij| is the same). We might be sensitive to this difference in some uses of the minimum image convention.

I have now modified so that |r_ij| in the shortest AND r_ij is chosen consistently amongst possible degenerate vectors.

This has increased computational cost for orthogonal cells.