Fixed NAN Bug
for non-periodic skyrmion_core_coordinates we find the coordinate by :

center_of_mass[0] = tube_center_of_mass_x[0] / double(num_core_spins);

However!!!

Just before the skyrmion is completely annihilated at the edges the num_core_spins is zero since the spins progressivelly rotate towards the possitive direction ; thus the cylinder is empty.

At that point the center of mass tries to do the calculation center_of_mass[0] = 0/0 which returns infinite or NAN. —> this results in accepting all trial_moves which cross the skyrmion_threshold, so we get this “sandwich-like” spin texture

To avoid this I introduced an if statement which in that case described returns zero ! this wont affect the biasing of the system since the \DeltaP will be also zero and at that point the Death_boundary will stop the simulation.

Note , the death_boundary before wasnt stopping the simulation since it wa never trigered from the trial spin getting accepted randomly throught the lattice