Cmake cleanup (#128)
* Remove _ROOT variables

We previously set the CMP0074 policy to NEW to avoid a warning about the use of _ROOT variables when finding packages.

The only affected package is FindMKL and we can avoid the issue simply by not using internal variables called _ROOT.

* Change cmake minimum required version in template

The cmake template for external projects which are added with the DownloadProject cmake mechanics was producing a warning due to the minimum version required being very old.

I've now update this to be the same minimum version as the rest of JAMS.

* Fix fftw3 assumed found if MKL not found

The code erroneously assumed fftw3 was found if MKL was not found. We need to check for this explicitly so we can fail if fftw3 is also not found.

* Use MKLROOT environment variable for MKL search

* Print found libraries list on new lines

* Give cmake error if dependency min version not found

* Cleanup JAMS version information

We now also print a warning if running in a dirty git repo. This should prompt novice users not to do serious calculations with uncommitted changes.

* Update readme