Find CUDA types from template parameters
In the cuSPARSE functions we now select the correct CUDA/cuSPARSE types based on the template parameters.

Note: In the `cusparseSpMV` function the compute type selection is more complicated in general as it depends on the value types of the matrix and the vectors. See https://docs.nvidia.com/cuda/cusparse/index.html#cusparse-generic-function-spmv