CUDA 11 support (#131)
* Set minimum CUDA version in cmake

In CUDA 11 the cuSparse cusparseDcsrmv() function has been removed. The replacement is the generic cusparseSpMV() function but this requires a lot of work to support the new function.

I've added code so we can now enforce a minimum CUDA version. Currently version 9 should still work with the old API.

* Update github actions

I've changed the matrix build so we now test CUDA 10 and 11. I've also fixed running the unit tests on the matrix build and removed unit tests from the quick build.

* Use cuSPARSE generic API for SpMV

In CUDA 11 the 'cusparse<X>csrmv' functions have been removed so we must use the new API. This involves quite a bit more book keeping.

I have retained the older API for backwards compatibility with CUDA 9/10.

* Find CUDA types from template parameters

In the cuSPARSE functions we now select the correct CUDA/cuSPARSE types based on the template parameters.

Note: In the `cusparseSpMV` function the compute type selection is more complicated in general as it depends on the value types of the matrix and the vectors. See https://docs.nvidia.com/cuda/cusparse/index.html#cusparse-generic-function-spmv