Fix lorentzian noise (#135)
* Improve num_freq and num_trunc docs

* Fix incorrect equation for bose-einstein noise

The equation was incorrect. The power spectrum should be (ℏω/kT) / (exp(ℏω/kT) - 1). I was missing a factor of kT which would cancel with the kT for the fluctuation dissipation theorem. This is also why the omega=0 expression looked very strange and was not the classical result it should have been.