Remove hard version number (#138)
* Remove hard coded version number

We don't actually use this version number, we get the version by inspecting the git log. So this value is both irrelavent and often out of date because it needs manual verison bumping.

* Warn if version number is unknown

Unknown versions are not traceable in the git history so shouldn't be used for production work. Warn the user.