Use sz component to weight skyrmion COM
In metadynamics we used to give all spins below a certain sz threshold an equal weighting for the center of mass calculation. Therefore only spins which cross this threshold can change the centre of mass. These are the spins on the edge of the skyrmion.

Testing at low temperature showed that this can have very non-physical effects, distorting the edge of the skyrmion and even breaking it before the skyrmion has really moved (i.e. the COM moves because of the distortion but the core stays in place).

Now we are weighting each spin in the COM calculation by Sz as an effective mass. This means that spins in the core have an effect (and a large effect) on moving the COM. If the core spins are pushed by the metadynamics then more of the skyrmion spins can be moved just by the regular Monte Carlo. In the end the skyrmion now moves much more coherently. This is very clear at low temperature and *should* fix issues with the energy landscape having a large peak in the initial position of the skyrmion.