Add setting to disable checking atom closeness (#140)
When using very large supercells (e.g. amorphous systems) the safety checking for the closeness of atoms in the motif can be very expensive.

This new setting allows the user to skip this check. It's a good idea to run the check in the first instance and then disable afterwards to avoid the large setup cost.