Merge metadynamics updates (#141)
* Fix mistake in interpolation formula.

* Add setting to toggle random spin selection in Monte Carlo

We used to always generate random indicies to choose a spin to move, but it may be beneficial to go in order, for example in metadynamics to ensure every spin is definately tried every MC step. So we add this as a toggle setting select_spins_by_random in the solver.

* Refactor metadynamics collective variables

Collective variables are now implemented as a very simple class which calculates a single value. The potential is now defined and manipulated in the MetadynamicsPotential class which holds the CollectiveVariable pointers.

Most CollectiveVariables are expensive so we have a derived caching class which helps to implement caching for Monte Carlo based solvers.

* Output collective variables as we insert gaussians

* Add setting to delay the start of metadynamics

This allows us to run normal monte carlo for some iterations to equilibrate before starting the metadynamics process.

* removing old cvar_bcs_

* HardBC in all cvars

* Use sz component to weight skyrmion COM

In metadynamics we used to give all spins below a certain sz threshold an equal weighting for the center of mass calculation. Therefore only spins which cross this threshold can change the centre of mass. These are the spins on the edge of the skyrmion.

Testing at low temperature showed that this can have very non-physical effects, distorting the edge of the skyrmion and even breaking it before the skyrmion has really moved (i.e. the COM moves because of the distortion but the core stays in place).

Now we are weighting each spin in the COM calculation by Sz as an effective mass. This means that spins in the core have an effect (and a large effect) on moving the COM. If the core spins are pushed by the metadynamics then more of the skyrmion spins can be moved just by the regular Monte Carlo. In the end the skyrmion now moves much more coherently. This is very clear at low temperature and *should* fix issues with the energy landscape having a large peak in the initial position of the skyrmion.

* Implement finite difference topological charge CV

This CV calculates the topological charge using finite differencing for the partial differentials in space.

* Warn users not to use topological_charge CV

* Clean metadynamics code

Co-authored-by: Giannis <pyic@leeds.ac.uk>