Refactor monitor convergence
Previously if a single monitor converged then the simulation would exit. But it's probably more sensible to only exit when all monitors have converged. However, this presents an issue that many monitors don't have a concept of convergence and so the true/false value of convergence is misleading and makes it difficult to check if all monitors are converged.

I've now refactored is_converged() to convergence_status() which will return a   Monitor::ConvergenceStatus enum which can be kDisabled, kConverged or kNotConverged. By default monitors will return kDisabled unless the base Monitor class virtual function is overriden. When the solver checks for convergence it will now check if any monitors are not converged or if all monitors are disabled.