Allow interaction checks to be toggled with config (#145)
The 'strict' interaction checks were always performed on interactions read in from the exchange Hamiltonian. However, it is sometimes desirable to disable a check, for example if we are only specifying a subset of exchange interactions in the unit cell with one Hamiltonian because another Hamiltonian will specify the rest.

We can now toggle each check on/off in the exchange settings. By default all checks are enabled.