Rename some vector operations
Some of the vector operations in cuda_device_vector_ops.h and for Vec types are poorly named to the point where some make no sense. For example abs is defined as ||a||^2 in cuda_device_vector_ops.h. All the uses for the functions seem to be correct, so likely a refactor has changed these in the past because they're not protected by namespaces.

I've now given several vector functions more verbose and clearer names as well as writing documentation for each function to be clear what is calculated.