Include read_potential
Can read a potential file for up to two collective variables.

Checks the size of potential with each of the cvars size.
Checks if a line is corrupted

Currently, I ignore the first line of the input file (I need to find a way to check if its titles or numbers  “string_is_comment(line)” are not working.

solver: {

module = "monte-carlo-metadynamics-cpu";

potential_file = "potential.tsv";

}