Kspace paths (#155)
* Fix kspace path displacement delta vector

The 'delta' vector is the step size to take on the discrete reciprocal space lattice when walking along a given path. Previously we normalized the components such that the delta of the component could be +/-1 or 0. This is incorrect for paths which are not 'simple', for example [0, 0, 0] -> [16, 24, 24] which would be given a delta of [1, 1, 1]. But this means we actually end up at [16,16,16]. Instead we need a delta of [2,3,3].

We fix this by using the greatest common divisor (gcd) of all of the elements in the vector to calculate the delta vector.

* Output total distance column for spectra

Added a column with the total distance along kspace of the path. This enables simple plotting of the x-axis with the correct distances.

* Fix extra bracket