Fix non-integer topological charge
There's a bug in the monitor for calculating the topological charge from the geometrical definition. We have a condition

    if (denominator <= 0.0) {
        	return 0.0;
    }

which used to be included to try and avoid 'exceptional' configurations in the metadynamics. However, when we are purely calculating the charge (not using the charge in the metadynamics) this condition is throwing out part of the sum over local Q contributions, hence we end up with non-integer Q. Removing this if statement means that the topological charge calculated with this monitor is now always integer.