Fix looping over stencil neighbours
Previously we looped over dx_indices_ as a proxy for all neighbouring sites in the finite difference stencil. However, if the dx and dy stencils are different then this won't cover all of the sites.

We now generalise this by using a vector of VectorSets to hold the unique indices for each site.