Allow Hamiltonians to be time dependent
Previously Hamiltonians could only be time dependent by accessing the global solver instance. Now we make time dependence an intrinsic part of the Hamiltonian, so all energy and field calculations now take a 'time' parameter. This decouples time dependent Hamiltonians from the global solver.

I have also change the way the time() method of solver works. We must now keep track of 'time_' in the solver to ensure it is correct during the mid point steps of numerical integration. Previously the time() was only updated at the end of each total step which is technically incorrect, so the midpoints would not have an updated time dependent field.