Restoring boundaries (#164)
* Restoring Boundaries

  collective_variables = (
    {
    name = "topological_charge_finite_diff";
    gaussian_width = 0.05;
    range_min = -1.5;
    range_max = 0.5;
    range_step = 0.01;
		lower_restoring_bc_threshold = -1.4;
		upper_restoring_bc_threshold = 0.4;
    },
  )
};

* Fix non-integer topological charge

There's a bug in the monitor for calculating the topological charge from the geometrical definition. We have a condition

    if (denominator <= 0.0) {
        	return 0.0;
    }

which used to be included to try and avoid 'exceptional' configurations in the metadynamics. However, when we are purely calculating the charge (not using the charge in the metadynamics) this condition is throwing out part of the sum over local Q contributions, hence we end up with non-integer Q. Removing this if statement means that the topological charge calculated with this monitor is now always integer.

* Fix typos

* Fix looping over stencil neighbours

Previously we looped over dx_indices_ as a proxy for all neighbouring sites in the finite difference stencil. However, if the dx and dy stencils are different then this won't cover all of the sites.

We now generalise this by using a vector of VectorSets to hold the unique indices for each site.

* Finite difference templates for square and hexagonal lattices

* Implement new stencils into monitor

* Fix restoring boundary conditions

* Fix incorrect use of current cvar value

Inside the metadynamics potential function we should always be using the cvar values passed into the function and never be using cvars_[n]->value(). The latter would be the same for both initial and trial moves, which is incorrect.

* Read restoring BC spring constant from config

* Fix cvars which are not intitialised

In principle any cvar beyond kMaxDimensions is never used, but they will be initialised with random data. Let's be safe and zero it out just in case.

* Fix array boundary error for restoring bcs

We're looking one off the end of the array when we lookup the potential at the boundary.

* Don't need to call cvars_[n]->value() multiple times

* Fix linspace producing extra point at the end

An incorrect floating point equality check meant we produced one extra point in the linspace which then causes issues at the boundaries.

* _metad_stat.tsv output correction

Metad_stats was outputting every time a gaussian was deposited when we are using tempered metadynamics, resulting at massive >600 MB files.

Now it outputs based on the output_steps from the config file like the potential and jams outputs.

* cvars.tsv file output steps output correction

Previously the cvars.tsv file created at the potential.cc was outputting in every trial move resulting in filesizes large than 500MB per simulation.

Now a new variable passed from the config file “cvars_output_steps = 1000; “ is introduced to control the output steps and file sizes.

Co-authored-by: Giannis <pyic@leeds.ac.uk>