Codebase cleanup (#163)
* Allow Hamiltonians to be time dependent

Previously Hamiltonians could only be time dependent by accessing the global solver instance. Now we make time dependence an intrinsic part of the Hamiltonian, so all energy and field calculations now take a 'time' parameter. This decouples time dependent Hamiltonians from the global solver.

I have also change the way the time() method of solver works. We must now keep track of 'time_' in the solver to ensure it is correct during the mid point steps of numerical integration. Previously the time() was only updated at the end of each total step which is technically incorrect, so the midpoints would not have an updated time dependent field.

* Refactor global namespace

Move all variables in globals.h inside the global namespace and remove 'using namespace global;' from all source code.

This makes it much clearer when we are using a global variable, especially when we are often shadowing variable names (such as the use of 'solver' in monitors).

* Fix dipole test Hamiltonian calls

* Change monitors to access solver as a reference

The pointer should never be null and we don't ever modify ownership of the pointer so this should be a reference. Ideally we will eventually turn in into a const reference (monitors should not be changing the solver!) but the Hamiltonians often need to be called by monitors.

* Cleanup includes

Cleanup the includes across base, hamiltonians, monitors.

* Initialize solver on construction

* Fix incorrect reading of timestep

* Remove using namespaces

* Fix missing namespaces after merge