Sinc bandwidth (#167)
* Redefine sinc pulse with bandwidth

The temporal_width is of course actually a frequency bandwidth. We redefine this and use the JAMS internal units of THz for frequencies.

* Add a sinc-cos temporal field

This allows the finite bandwidth sinc function to be centered in frequnecy space, filtering out a limited region of frequencies. See Träger, Sci. Rep. 10, 18146 (2020) [https://doi.org/10.1038/s41598-020-74785-4]