Fix dipole tests to clean globals
We should be cleaning and resetting all global variables between tests to start from a fresh slate. We now do this in the tear down. We also do a cudaDeviceReset() at SetUp to ensure the device is in a clean state. Note that's important where we call cudaDeviceReset(). We should not do it in TearDown() because this can break any classes which have not yet called their destructors but contain pointers to cuda allocated memory.