Fix dipole segfault (#170)
* Change 2 atom dipole test moments

Using 2.0 and 1.0 leaves the possibility for coincidence. Changing to numbers which are not multiples of each other could uncover issues.

* Make SyncedMemory copy operator argument const ref

Avoids a copy, which was previously used for the copy-swap idiom. Also ensures that the copy and move operators are not equivalent.

* Give CUDA error detail from SyncedMemory

* Remove copy-swap idiom from 1D MultiArray

This was already done in the template class but not in the 1D specialisation.

* Fix CUDA library linking for tests

This was broken by a previous refactor and hasn't been tested since.

* Move all dipole tests into single test file

* Fix segfaults in CudaDipoleFFTHamiltonian

For some reason we are passing host pointers where we should be passing device pointers. It's not clear why this hasn't been breaking this earlier. Presumably because we use this Hamiltonian so infrequenctly and the CUDA tests were disabled.

* Fix incorrect factors of mus

The magnetic moment 'mus' must appear for both sites 'i' and 'j' in the field contributions. Remembering that Hᵢ = (1/μᵢ)(∂ℋ/∂Sᵢ) so the dipole Hamiltonian should only be derived with respect to Sᵢ because the 1/μᵢ factor is accounted for inside the solver.

* Fix incorrect indexing for mus

Here 'i' refers to a reciprocal space unit cell index, not the spin number which is instead 'index'.

* Fix CUDA dipole bruteforce cutoff

To be consistent with the other dipole Hamiltonians, and to guard against spin inclusion on the edge of the cutoff sphere we should do proper floating point checks.

* Add swap method to MultiArray

Allows us to force deallocate

* Fix dipole tests to clean globals

We should be cleaning and resetting all global variables between tests to start from a fresh slate. We now do this in the tear down. We also do a cudaDeviceReset() at SetUp to ensure the device is in a clean state. Note that's important where we call cudaDeviceReset(). We should not do it in TearDown() because this can break any classes which have not yet called their destructors but contain pointers to cuda allocated memory.

* Fix incorrect size of indicies array in test

* Add override annotation

* Change gh checkout action to v3

Previous version v2 is now deprecated.