Only include and link cudart not cuda
We need to ensure we only use the cuda runtime api (libcudart) and not the low level cuda calls (libcuda). This is because libcuda only comes with the driver, not the runtime api, so we cannot find libcuda on systems without the driver installed (e.g. ARC head nodes and CPU nodes) even though the cuda runtime software is installed.