Only include and link cudart not cuda (#174)
* Only include and link cudart not cuda

We need to ensure we only use the cuda runtime api (libcudart) and not the low level cuda calls (libcuda). This is because libcuda only comes with the driver, not the runtime api, so we cannot find libcuda on systems without the driver installed (e.g. ARC head nodes and CPU nodes) even though the cuda runtime software is installed.

* Fix missing host_cuda_malloc_ from move and swap

Without moving and swapping the host_cuda_malloc_ variable the class will use the wrong functions to free the memory.