Fix interaction calculator search radius
To calculate the interactions within a given radius we need to search all unitcells which can be generated within that radius. For very skew cells this can require a very large number of cells even for a very short interaction radius.

Previously the calculation of the number of cells along each lattice vector to search in was incorrect. It used the assumption that we only needed the length along each vector. Actually we need to know the 'height' of the parallelpiped in each direction. The same logic applies in the searches for the minimum image convention where this was implemented correctly some time ago. I've used the same logic as the minimum image convention here and the interaction calculator results are now correct even for very skew cells.