Fix invalid array lookup if CV is exactly min_range
The logic previously meant that if the CV was exactly on the bottom of the allowed range the lower_index for cvar_sample_coordinates_ was found to be -1 (with upper_index 0) causing the array accesses to give junk and then break the interpolation.

Because the std::lower_bound function always returns a pointer to the first value larger than the search value this can return the first element in the array in which case we cannot use -1 trivially to find the lower index.

Now, if lower_bound finds the first element we make the lower and upper index the same.