Fix spin current units
The calculation is in JAMS internal units. We want to convert to SI for output. Js should have dimensions of m/s. In JAMS lengths are in terms of the lattice constant and times are in picoseconds so we need to multiply by the lattice constant and convert 1/ps to 1/s.