Fix lattice volume calculation (#185)
Previously, the output file volume was incorrect. It was printed under the 'unitcell' heading, but the code indicatated it was the supercell volume, but the printing happens before the supercell is generated. So in the end it was always the volume of a cubic unit cell.

I've removed the volume() method from the lattice class to avoid any ambiguity about what this volume corresponds to. Now in the unit cell and lattice sections I output the correct volumes.