Add setting to choose spin normalisation at setup
We previously always normalise spins to unit vectors. Now we have the GSE solver which does not assume unit length it makes sense not to always normalise the spin length. A setting is now added to the lattice settings block to toggle the normalisation. The default behaviour is retained that spins should be normalised.