Fix spin pumping monitor for all solvers
The spin pumping monitor needs the differential ds/dt. For the Heun solver this is simply taken from the Euler step, however for solvers such as the Runge-Kutta 4 there's not a clear single value to use for the derivative and the global ds_dt variable was not being updated, leading to empty output.

Now we calculate ds/dt in the spin pumping monitor using a simple backward difference calculation. I generalised the monitor class so the is_updating method is virtual. The spin pumping monitor can then use this to grab a copy of the spin array one step before the point at which it will actually update.