Support induced moments (#188)
* Add general spin equation (GSE) solver

This solver is based on P. Ma et al., Longitudinal magnetic fluctuations in Langevin spin dynamics, Phys. Rev. B 86, 054416 (2012) (http://dx.doi.org/10.1103/PhysRevB.86.054416) and M.O.A. Ellis et al., Role of longitudinal fluctuations in L1₀FePt, Phys. Rev. B 100, 214434 (2019) (https://doi.org/10.1103/PhysRevB.100.214434).

The GSE allows the spin in the Landau-Lifshitz equation not to be of unit length. It requires a Landau energy term to stabilise the equilibrium length of the spin.

* Add CUDA Landau Hamiltonian

This implements a term like

$H = \sum_i A_i|S_i|^2 + B_i|S_i|^4 + C_i|S_i|^6$

where the energy depends on the length of the spin. This Hamiltonian term should only be used with the GSE solver. Other solvers keep the spin unit length and this term will be a constant energy.

* Fix missing gyro factor

* Initial implementation of biquadratic for testing

* Add biquadratic exchange

* Output magnetic moments to h5

* Add induced-spin-pulse physics module

Adds a new physics module to simulate physics where the length of atomic moments is changed by laser excitations. For example the OISTR process [Dewhurst, Nano Lett. 18, 1842 (2018)](https://dx.doi.org/10.1021/acs.nanolett.7b05118). It does not implement a detailed physical model nor does it transfer spin moment between sites. It simply adjust the spin lengths.

* Add setting to choose spin normalisation at setup

We previously always normalise spins to unit vectors. Now we have the GSE solver which does not assume unit length it makes sense not to always normalise the spin length. A setting is now added to the lattice settings block to toggle the normalisation. The default behaviour is retained that spins should be normalised.

* Add triple q initializer

Allows complex alternating spin textures such as antiferromagnetic states to be initialised even when all atoms have the same material name.

* Add coherent induced spin pulse

This allows the induced spin pulse to grow the spins in a specified polarisation direction rather than each spins current direction.

* Fix spin pumping monitor for all solvers

The spin pumping monitor needs the differential ds/dt. For the Heun solver this is simply taken from the Euler step, however for solvers such as the Runge-Kutta 4 there's not a clear single value to use for the derivative and the global ds_dt variable was not being updated, leading to empty output.

Now we calculate ds/dt in the spin pumping monitor using a simple backward difference calculation. I generalised the monitor class so the is_updating method is virtual. The spin pumping monitor can then use this to grab a copy of the spin array one step before the point at which it will actually update.

* Update pre-merge.yml

* Update pre-merge.yml

* Fix biquadratic-exchange is CUDA only

* Fix settings for sparse matrix checks