Fix interaction symmetry operations (#200)
When we check local symmetry operations in the dipole methods or when we generate symmetric interactions in the exchange interaction we were using the rotation matricies of the symmetry operations of the whole crystal. This is the crystal point group. In many cases the point group of the Wyckoff positions we are using is the same as the point group of the crystal. But this is not always the case.

This meant that sometimes our symmetry checks or use of symops to generate exchange interactions, would fail. Generally the fall back then has been to turn off the symmetry operations and avoid checks and write interactions by hand.

The correct solution is actually to calculate the local point group of each site in the unit cell motif and then use these to check/generate interactions. The local point group can be found by considering the rotational symmetry operations of the crystal which leave that point invariant.

This is now implemented and the local point group operations can be returned with atom_motif_local_point_group_symops().