Fix unitcell-average monitor (#201)
* Fix default output path

If no output path is given on the command line, then some outputs are written with a leading "/" which is to the root file system which is not writable.

We now using "." as the default path

* Fix missing '.h5' in file reference

The magnetisation data reference in the xdmf has the the file extension missing. This breaks the use of the xdmf in paraview.