Fix unitcell average monitor (#202)
* Fix default output path

If no output path is given on the command line, then some outputs are written with a leading "/" which is to the root file system which is not writable.

We now using "." as the default path

* Fix missing '.h5' in file reference

The magnetisation data reference in the xdmf has the the file extension missing. This breaks the use of the xdmf in paraview.

* Fix time units in h5 output

Internal times are now in picoseconds, so we don't need to divide through in the h5 output.

* Fix unitcell-average position output

We should be consistent between the h5 and unitcell-average output such that they can be overlaid in paraview if needed.

I've chosen to output in nanometers for both. This makes any measurement of distances in paraview meaningful.

* Write more detail about units

Write details about what units are being used in h5 files for clarity.