Fix psd output (#207)
* Fine tuning the truncation algorithm

The magnitude of the memory kernel can be quite different depending on the lorentzian parameters, so we should make the detection of zero relative to the magnitude of the memory.

* Fix PSD output

The output was incorrectly using the original delta_omega rather than the updated version after choosing a the resolution. I have fixed this and also now output the PSD data itself, rather than just the function call.