Fix k-space path raster algorithm
Previously, if a path didn't have a small greatest common denominator then the rastering of the path could simply jump from the start to the end point, meaning that no points are sampled on the path of interest. An example of this is along the K-M path of a hexagonal system e.g.

hkl_path = (
      [ 0.0, 0.0, 0.0],               // G
      [ 0.333333, 0.333333, 0.0], // K
      [0.0, 0.5, 0.0],                // M
      [0.0, 0.0, 0.0],                // G
      [0.0, 0.0, 0.5]);               // A

Now I have implemented a 3D version of Bresenham's line algorithm which finds the nearest points to the path rather than only considering points exactly on the path.