Fix local point group calculation
We previously found the local point group by using only the rotation operations in the space group, simply ignoring the translation operations. However, this is incorrect. Any element with a translation should be completely ignored for determining the point group.

This is now fixed by skipping any symmetry operation which includes a translation.