Fix exchange symops only applied to JAMS format
For reasons that are not clear (nor correct) the symops option only works if the exchange input is in JAMS format rather than KKR format. By this point the interactons are already in a consistent format in terms of cartsian coordinates, so we can apply the symmetry operations to them regardless of their original format.