Merge crystal field (#219)
* Fix exchange symmetry operations

After moving to using local symmetry operations on each site some exchange interactions were broken. There were two issues:

1) No symmetry operations were generated when specifying with materials because the unit cell index was -1 (interaction data was completed after symops)

2) After applying the rotations I was no translating back into the unit cell to identify if it was the same atom.

* Refactor crystal field generator notebook

I've changed the normalisation to include the Sqrt[4 Pi/(2 l + 1)] in the notebook and also to use |m| for the tesseral harmonics.

* Update crystal field energy and field calculations

Using the change in normalisation from the notebook and correcting the form of the equation used in comparison to Chris

* Fix ComplexExpand for Re and Im

* Dociment crystal field

* Fix k-space path raster algorithm

Previously, if a path didn't have a small greatest common denominator then the rastering of the path could simply jump from the start to the end point, meaning that no points are sampled on the path of interest. An example of this is along the K-M path of a hexagonal system e.g.

hkl_path = (
      [ 0.0, 0.0, 0.0],               // G
      [ 0.333333, 0.333333, 0.0], // K
      [0.0, 0.5, 0.0],                // M
      [0.0, 0.0, 0.0],                // G
      [0.0, 0.0, 0.5]);               // A

Now I have implemented a 3D version of Bresenham's line algorithm which finds the nearest points to the path rather than only considering points exactly on the path.

* Fix local point group calculation

We previously found the local point group by using only the rotation operations in the space group, simply ignoring the translation operations. However, this is incorrect. Any element with a translation should be completely ignored for determining the point group.

This is now fixed by skipping any symmetry operation which includes a translation.

* Fix std::exception inheritance

By default `std::exception` was inherited as private which then meant `GeneralException` and all derived classes could not be caught using `catch (const std::exception& ex)`

* Fix exchange symops only applied to JAMS format

For reasons that are not clear (nor correct) the symops option only works if the exchange input is in JAMS format rather than KKR format. By this point the interactons are already in a consistent format in terms of cartsian coordinates, so we can apply the symmetry operations to them regardless of their original format.

* Allow don't gitignore cfg in test directory

* Add exchange symops integration test

Starting with a hard test for the quite low symmetry system CrPS4. The local point group symmetry operations must be found correctly for the Cr sites because the crystal point group would give non-existent interactions due to the inversion operation.

* Change hdf5 linkage to dynamic