Generalise pinned boundaries (#226)
* Generalise pinned boundaries

Pinned boundaries can now be defined on any face of the supercell.

* Generalise Bloch domain wall initialiser

We can now specify the direction and orientation of the domain wall. The initialiser also now rotates the spins with respect to the reference domain wall state. This means any existing system (e.g. ferrimagnetic, antiferro, thermalised etc...) is maintained, with the domain wall superimposed.