Move check_interaction_list_symmetry
It was previously inside the post_process_interactions, but if we use this for th topological CV it will fail the check because we only use if for the template, so there is no symmetric interaction (j -> i). It makes more sense to move this into the neighbour list generation functions. Moving it there also does not change the logic of the rest of JAMS.