Implement file parallel multi walker metadynamics
We use a hdf5 file to synchronise the potential between multiple walkers.

The parallel file is specified with the config setting `parallel_shared_potential_file` and the steps between sychronisation with `parallel_shared_potential_sync_steps`.

We use some basic file locking with an attribute in the hdf5 file to ensure only one process reads and writes to the file at a time.

For the creation of the hdf5 file we use posix to ensure only a single process can create it. Of course for each new parallel simulation we must delete any pre-exisiting shared file outside of JAMS.