Feature multiple walkers (#233)
* Make post_process_interactions accessible

We need to use this in the topological charge CV so it needs to be in the header.

* Move check_interaction_list_symmetry

It was previously inside the post_process_interactions, but if we use this for th topological CV it will fail the check because we only use if for the template, so there is no symmetric interaction (j -> i). It makes more sense to move this into the neighbour list generation functions. Moving it there also does not change the logic of the rest of JAMS.

* Use post_process_interactions for topological charge CV

We need the lattice offsets to be calculate, the simplest way is to post process the interactions to fill the missing information.

* Implement file parallel multi walker metadynamics

We use a hdf5 file to synchronise the potential between multiple walkers.

The parallel file is specified with the config setting `parallel_shared_potential_file` and the steps between sychronisation with `parallel_shared_potential_sync_steps`.

We use some basic file locking with an attribute in the hdf5 file to ensure only one process reads and writes to the file at a time.

For the creation of the hdf5 file we use posix to ensure only a single process can create it. Of course for each new parallel simulation we must delete any pre-exisiting shared file outside of JAMS.

* Change to using a lock file

On the cluster the attibute locking of the hdf5 file doesn't work. I think the I/O is a bit slower than my macbook and so multiple processes do try to open the hdf5 file at the same time and crash when they can't.

* Fix file locking

It seems that some processes were still trying to access the hdf5 while it was open with another process. I think it's because the file object is not closed until it goes out of scope (HighFive has no explicit close method).