Fix/restoring boundaries (#234)
* Return early when outside of restoring boundary

An optimisation to avoid unnessecary calculations of gaussians which will be multiplied by zero.

* Fix restoring boundaries

Previously the restoring boundaries were using the value of the potential at the top left or bottom right corner as the constant offset. This meant the system could still get stuck at the boundary. This mistake is probably left over from when we assumed we would use this boundary only for 1D problems.

Now we correctly use the value of the potential nearest to the boundary as the offset so that the potential smoothly becomes harmonic across the boundary.