Fix resize arrays before initial random numbers
The array resizing was being done after the initial generation of random numbers. This meant curandGenerateNormalDouble was given a size of zero and no initial random numbers would be generated.

This is probably ok, because the arrays will get filled during the warm up period, but it is not designed correctly.