CudaThermostatQuantumSpde minor fixes (#242)
* Fix resize arrays before initial random numbers

The array resizing was being done after the initial generation of random numbers. This meant curandGenerateNormalDouble was given a size of zero and no initial random numbers would be generated.

This is probably ok, because the arrays will get filled during the warm up period, but it is not designed correctly.

* Fix generated eta0 even if no zero point

We were generating noise into the eta0 array even if it is not initialized and unneeded because  do_zero_point_ is false.

We now only generate if needed.

* Fix warmup steps +1 bug

The order of these lines is wrong, such that the warmup ends up happening for `num_warm_up_steps_ + 1` times, rather than `num_warm_up_steps_` times.

* Fix remove duplicate loop

The duplicated loop has no effect (and is probably optimised out).

* Fix incorrect pointer size

Shouldn't cause any issue because it decays to a pointer, but is inconsistent when A, eta, z are of size 4.

* Simplify warmup

* Annotate with override

* Add inline specifier to kernels